Revert "Logo fix"
