+++ workout-tracker/src/App.css
@@ -1,6 +1,6 @@
 .App {
   text-align: center;
 }
-/* * {
+* {
   margin: 0;
-} */
+}
