+++ workout-tracker/src/components/Backdrop/Backdrop.css
@@ -5,5 +5,5 @@
   z-index: 100;
   left: 0;
   top: 0;
-  background-color:#5c0d91;
+  background-color: rgba(0, 0, 0, 0.5);
 }
