+++ workout-tracker/src/components/Layout/Layout.js
@@ -37,7 +37,7 @@ const StyledContainer = styled.section`
   }
   .ant-layout-sider {
     padding: 0.5rem;
-    background-color: #5c0d91;
+    background-color: #001529;
     .ant-layout-content {
       margin: 0;
       padding: 0;
