+++ workout-tracker/src/components/Logo/Logo.css
@@ -1,4 +1,3 @@
-@import url('https://fonts.googleapis.com/css?family=Viga&display=swap');
 .logo {
   font-size: 1rem;
   font-weight: bold;
@@ -14,18 +13,4 @@
 .logo-container {
   flex: 1;
   text-align: center;
-}
-
-
-#newstuff {
-  font-size: 3.5em;
-  color: #0086c9;
-  /* line-height: 0.8em; */
-  text-align: left;
-  margin: 1em;
-  font-family: 'Viga', sans-serif;
-}
-h3 span {
-  color: #fff;
-  font-family: 'Viga', sans-serif;
 }
