+++ workout-tracker/src/components/Logo/Logo.js
@@ -3,11 +3,6 @@ import logoImage from '../../assets/images/logo-svg-white.svg'
 import './Logo.css';
 import { Link } from 'react-router-dom';
 
-const logo = props => <div className="logo-container">
-    <Link to='/workouts'>
-    <h3 id="newstuff">
-            Be<span>Fit</span>
-          </h3>
-    </Link></div>
+const logo = props => <div className="logo-container"><Link to='/workouts'><img className="image" src={logoImage} alt="logo" /></Link></div>
 
 export default logo;
