+++ workout-tracker/src/components/MainNavBar/SideNavigation/DashboardNavItem.css
@@ -9,13 +9,7 @@
   text-align: left;
   font-size: 1.5rem;
   color: #696969;
-  border-left: 11px solid #5c0d91;
-}
-.ant-layout-header{
-  background-color: #5c0d91;
-}
-.ant-layout-sider-zero-width-trigger {
- background: #5c0d91;
+  border-left: 11px solid #001529;
 }
 
 .dashboard-list-item:hover,
