+++ workout-tracker/src/views/WorkoutSession/WorkoutSession.js
@@ -1,7 +1,8 @@
-import React from 'react';
-import Watch from '../../components/Watch/Watch';
-import { connect } from 'react-redux';
+import React from "react";
+import Watch from "../../components/Watch/Watch";
+import { connect } from "react-redux";
 import {
+  PageHeader,
   Statistic,
   Row,
   Col,
@@ -10,22 +11,64 @@ import {
   List,
   Button,
   Modal,
+  Carousel,
   Alert
-} from 'antd';
+} from "antd";
 import {
   chooseExercise,
   finishExercise,
   endWorkout
-} from '../../store/actions/workoutsActions';
-import styled from 'styled-components';
+} from "../../store/actions/workoutsActions";
+import styled from "styled-components";
 
-const StyledWorkoutSession = styled.div``;
+const StyledWorkoutSession = styled.div`
+  /* font-size: .8rem; */
+
+  line-height: 1;
+
+  margin: 0 auto;
+
+  .top {
+    display: flex;
+  }
+
+  .exercise-picture {
+    width: 50%;
+  }
+
+  .text {
+    width: 50%;
+  }
+
+  .picture-text {
+    display: flex;
+    width: 66%;
+  }
+
+  img {
+    width: 100%;
+  }
+
+  .exercise {
+    cursor: pointer;
+  }
+
+  @media only screen and (max-width: 768px) {
+    .carousel {
+      display: none;
+    }
+  }
+  @media only screen and (min-width: 768px) {
+    .cover {
+      display: none;
+    }
+  }
+`;
 
 class WorkoutSession extends React.Component {
   componentDidMount = () => {
-    // const startButton = document.querySelector(".btn-start");
-    // startButton.click();
-    // location.reload();
+    const startButton = document.querySelector(".btn-start");
+    startButton.click();
   };
 
   endWorkout = () => {
@@ -42,13 +85,13 @@ class WorkoutSession extends React.Component {
     });
   };
 
-  handleCancel = (e) => {
+  handleCancel = e => {
     this.setState({
       visible: false
     });
   };
 
-  handleOk = (e) => {
+  handleOk = e => {
     this.setState({
       visible: false
     });
@@ -65,82 +108,110 @@ class WorkoutSession extends React.Component {
   render() {
     return (
       <StyledWorkoutSession>
-        <Card
-          style={{
-            fontSize: 14,
-            fontColor: 'white',
-            backgroundColor: '#5c0d91'
-          }}
-        >
-          <Card //Top Card with picture / watch  and Details reps etc
-            // type="inner"
-            bordered={false}
-          >
-            <div style={{}}>
-              <Card
-                bordered={false}
-                style={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  justifyContent: 'center'
-                }}
-                title={this.props.currentExercise[0].exercise_name}
-              >
-                <div
-                  style={{
-                    height: '350px',
-                    display: 'flex',
-                    flexDirection: 'row',
-                    justifyContent: 'center'
-                  }}
-                >
-                  {/* <img
-                    bordered={false}
-                    // style={{ width: 500 }}
-                    alt="Exercise explanation"
-                    src={this.props.currentExercise[0].picture_one}
-                  /> */}
-
-                  <video autoplay loop playsinline muted controls width="90%" height="auto">
-                    <source
-                      alt="Exercise explanation"
-                      src={this.props.currentExercise[0].video}
-                      type="video/mp4"
+        <Row type="flex" justify="space-around">
+          <Col span={12}></Col>
+          <Col span={12}></Col>
+          <PageHeader
+            onBack={() => window.history.back()}
+            title={
+              this.props.myWorkout ? this.props.myWorkout.workout_name : null
+            }
+          />
+          <Watch />
+        </Row>
+        {this.props.currentExercise ? (
+          <>
+            <Row type="flex" justify="center" align="top">
+              <Col md={7}>
+                {this.props.allExercises ? (
+                  <div>
+                    {/* {} */}
+                    <List
+                      size="small"
+                      header={<h3>Choose Exercises</h3>}
+                      bordered
+                      dataSource={this.props.allExercises
+                        .reduce((acc, current) => {
+                          const x = acc.find(
+                            item => item.exercise_name === current.exercise_name
+                          );
+                          if (!x) {
+                            return acc.concat([current]);
+                          } else {
+                            return acc;
+                          }
+                        }, [])
+                        .map(exercise => exercise.exercise_name)}
+                      renderItem={item => (
+                        <List.Item
+                          onClick={() => this.props.chooseExercise(item)}
+                        >
+                          {<Button>{item}</Button>}
+                        </List.Item>
+                      )}
                     />
-                    Your browser does not support the video tag.
-                  </video>
-                </div>
-              </Card>
-
-              <Card
-                title="stats"
-                bordered={false}
-                style={{ display: 'flex', flexDirection: 'column' }}
-              >
+                  </div>
+                ) : (
+                  <p className="button" onClick={this.endWorkout}>
+                    Finish Workout
+                  </p>
+                )}
+              </Col>
+              <Col sm={7}>
                 <Card
+                  cover={
+                    <>
+                      <div className="carousel">
+                        <Carousel
+                          dots={false}
+                          effect="fade"
+                          autoplay
+                          autoplaySpeed={1000}
+                        >
+                          <div>
+                            <img
+                              alt="Exercise explanation"
+                              src={this.props.currentExercise[0].picture_one}
+                            />
+                          </div>
+                          <div>
+                            <img
+                              alt="Exercise explanation"
+                              src={this.props.currentExercise[0].picture_two}
+                            />
+                          </div>
+                        </Carousel>
+                      </div>
+                      <div className="cover">
+                        <img
+                          alt="Exercise explanation"
+                          src={this.props.currentExercise[0].picture_one}
+                        />
+                      </div>
+                    </>
+                  }
                   title={this.props.currentExercise[0].exercise_name}
-                  bordered={false}
                   actions={[
                     <Statistic
                       title="Sets to complete"
-                      prefix={<Icon type="list" />}
-                      style={{ cursor: 'default' }}
+                      prefix={<Icon type="unordered-list" />}
+                      style={{ cursor: "default" }}
                       value={this.props.currentExercise.length}
                     />,
                     <Statistic
                       title={
                         this.props.currentExercise[0].reps
-                          ? 'Repetitions'
-                          : 'Duration'
+                          ? "Repetitions"
+                          : "Duration"
                       }
-                      prefix={<Icon type="sync" />}
+                      prefix={<Icon type="sync" spin />}
                       value={
                         this.props.currentExercise[0].reps ||
                         (this.props.currentExercise[0].duration
                           ? this.props.currentExercise[0].duration
-                          : '20seconds')
+                          : "20seconds")
                       }
-                      style={{ cursor: 'default' }}
+                      style={{ cursor: "default" }}
                     />,
                     <Statistic
                       title="Next Exercise"
@@ -151,10 +222,10 @@ class WorkoutSession extends React.Component {
                               this.props.currentExercise[0].id
                             )
                           }
-                          type="fast-forward"
+                          type="double-right"
                         />
                       }
-                      style={{ cursor: 'default' }}
+                      style={{ cursor: "default" }}
                       value=" "
                     />
                   ]}
@@ -163,49 +234,16 @@ class WorkoutSession extends React.Component {
                     description={`Equipment Needed: ${this.props.currentExercise[0].equipment}`}
                   />
                 </Card>
-
-                <div
-                  style={{
-                    margin: 20,
-                    display: 'flex',
-                    justifyContent: 'center'
-                  }}
-                >
-                  <Button type="danger" onClick={this.endWorkout}>
-                    Finish Workout
-                  </Button>
-                </div>
-              </Card>
-
-              <Card
-                title="Timer"
-                bordered={false}
-                style={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  alignContent: 'spaceAround'
-                }}
-              >
-                <Watch />
-              </Card>
-            </div>
-          </Card>
-          <Card //Bottom Card with Exer & Instructions
-            style={{ marginTop: 16 }}
-            type="outer"
-          >
-            <div
-              style={{
-                background: '#ECECEC',
-                display: 'flex',
-                flexDirection: 'row'
-              }}
-            >
-              <Card //Instructions bar
-                title="Instructions"
-                bordered={false}
-              >
-                <Card bordered={false}>
+              </Col>
+              <Col md={7}>
+                {this.props.currentExercise ? (
+                  <div style={{ marginTop: "1rem" }}>
+                    <Button type="primary" onClick={this.showModal}>
+                      Video Instruction
+                    </Button>
+                  </div>
+                ) : null}
+                <Card bordered={false} style={{ lineHeight: 1.2 }}>
                   <Alert
                     message="Instructions"
                     description={this.props.currentExercise[0].description}
@@ -213,47 +251,38 @@ class WorkoutSession extends React.Component {
                   />
                   {/* {`${this.state.initial} ==> ${this.props.currentExercise.length} ===> ${this.props.allExercises.length}`} */}
                 </Card>
-              </Card>
-
-              <Card // Excercise List
-                title="Excercise List"
-                bordered={false}
-              >
-                <List
-                  size="small"
-                  header={<h3>Choose Exercises</h3>}
-                  dataSource={this.props.allExercises
-                    .reduce((acc, current) => {
-                      const x = acc.find(
-                        (item) => item.exercise_name === current.exercise_name
-                      );
-                      if (!x) {
-                        return acc.concat([current]);
-                      } else {
-                        return acc;
-                      }
-                    }, [])
-                    .map((exercise) => exercise.exercise_name)}
-                  renderItem={(item) => (
-                    <List.Item onClick={() => this.props.chooseExercise(item)}>
-                      {
-                        <Button style={{ textAlign: 'left' }} type="link" block>
-                          {item}
-                        </Button>
-                      }
-                    </List.Item>
-                  )}
+              </Col>
+            </Row>
+          </>
+        ) : (
+          <Button type="primary" onClick={this.endWorkout}>
+            End Workout
+          </Button>
+        )}
+        {this.props.currentExercise ? (
+          <Modal
+            title={this.props.currentExercise[0].exercise_name}
+            visible={this.state.visible}
+            onCancel={this.handleCancel}
+            onOk={this.handleOk}
+          >
+            {this.state.visible && (
+              <video width="100%" height="auto" autoPlay controls>
+                <source
+                  src={this.props.currentExercise[0].video}
+                  type="video/mp4"
                 />
-              </Card>
-            </div>
-          </Card>
-        </Card>
+                Your browser does not support the video tag.
+              </video>
+            )}
+          </Modal>
+        ) : null}
       </StyledWorkoutSession>
     );
   }
 }
 
-const mapStateToProps = (state) => {
+const mapStateToProps = state => {
   return {
     allExercises: state.workouts.allExercises,
     currentExercise: state.workouts.currentExercise,
